requirements.txt
torch>=2.2.0
numpy>=1.24.0
PyYAML>=6.0.1
optuna>=3.5.0
tqdm>=4.66.0
flwr>=1.7.0
codecarbon>=2.4.0
black>=24.3.0
ruff>=0.4.2
rich>=13.7.1
textual>=0.58.0
openai>=1.35.7
anthropic>=0.34.2
google-generativeai>=0.7.2
mistralai>=0.4.2
requests>=2.32.3
